+++ NEWS.md
@@ -8,6 +8,8 @@ Unversioned
   * Deprecate ``load_model_only`` keyword argument to ``solve``.
   * Add groups of constraints with ``@addConstraints`` macro.
   * Unicode operators now supported, including ``∑`` for ``sum``, ``∏`` for ``prod``, and ``≤``/``≥`` (requires Julia 0.3)
+  * Quadratic constraints supported in ``@addConstraint`` macro.
+  * Quadratic objective supported in ``@setObjective`` macro.
 
 Version 0.5.4 (June 19, 2014)
 ----------------------------